requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
pydantic==2.6.1
groq==0.9.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-dotenv==1.0.1
streamlit>=1.35.0
pandas>=2.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
